Update all source versions when updating the app
Currently, when an app is updated, sources that existed in the app before are **not** updated automatically. This behavior is desired to avoid an app spec change that only changes, for example the scale of a component from picking up new code.

By setting the `UpdateAllSourceVersions` flag, the action now forces updates to **all** the sources of the app (either git based or image based) on the update.