Extract PR number from event instead of using RefName (#145)
On normal pull-request events, the `RefName` attribute is usually `$prNumber/merge`. However, when **merging** a pull-request, it becomes the name of the branch the PR is merged into. This beraks the connection between creating and deleting preview apps.

This changes the method of figuring out the underlying pull-request by looking at the event payload directly and parsing the respective number out of there.